CN-7312: moving user-util to separate folder
