+++ app/components/atoms/InterestRateInfo/InterestRateInfo.js
@@ -11,7 +11,7 @@ import Card from "../Card/Card";
 import CoinIcon from "../CoinIcon/CoinIcon";
 import cryptoUtil from "../../../utils/crypto-util";
 import RateInfoCard from "../../molecules/RateInfoCard/RateInfoCard";
-import { isUSCitizen } from "../../../utils/user-util";
+import { isUSCitizen } from "../../../utils/user-util/user-util";
 import { getColor } from "../../../utils/styles-util";
 import { COLOR_KEYS } from "../../../constants/COLORS";
