+++ app/components/screens/Home/Home.js
@@ -3,7 +3,7 @@ import { connect } from "react-redux";
 import { bindActionCreators } from "redux";
 import * as appActions from "../../../redux/actions";
 import mixpanelAnalytics from "../../../utils/mixpanel-analytics";
-import { isKYCRejectedForever } from "../../../utils/user-util";
+import { isKYCRejectedForever } from "../../../utils/user-util/user-util";
 import API from "../../../constants/API";
 import apiUtil from "../../../utils/api-util";
 import SplashScreen from "../SplashScreen/SplashScreen";
