+++ app/redux/user/userReducer.js
@@ -1,5 +1,5 @@
 import ACTIONS from "../../constants/ACTIONS";
-import { mapProfile } from "../../utils/user-util";
+import { mapProfile } from "../../utils/user-util/user-util";
 import { THEMES } from "../../constants/UI";
 
 const initialState = {
